core/standards/scripts/scheduler2/location_hash/204.html

<!doctype html>
<title>Location hash - multiple set hash during window.onload 2</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var histlength;
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
    histlength = history.length;
});
</script>
<script>
window.onload = function () {
    t.step(function() {
        var i = 50;
        do {
            location.hash = '#foo' + i;
            location.hash = '#bar' + i;
        } while (--i);
    });
    t.step(function() {
        assert_equals( location.hash, '#bar' + 1 );
        assert_equals( history.length, histlength );
    });
    t.done();
}
</script>
<div id="log"></div>
